requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pypdf2==3.0.1
python-docx==1.1.0
spacy==3.7.2
streamlit==1.28.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.27.0
ollama==0.1.8
